EmbeddedDXIL and SPIR-V options
The options, '-embed-dxil' and '-embed-spirv' in slangc, cause
a target dxil or spirv to be compiled and stored in the
translation unit IR when written to a slang-module. Subsequent
changes intend to load the blobs from slang-modules when loaded
as to short-circuit additional processing and pass through
directly to DXC or Spirv-tools respectively, for linking.