Set debug working directory for slangc.exe
This commit changes the working directory setting in Visual Studio
Debugging from "build" directory to the root of the repository.

It is often the case that the users debug slang shaders under "tests"
directory, and the working directory needs to be manually changed to
the root directory of the repository.