Error out when constructing tensor views from tensors with 0 stride.
This avoids a problem with broadcasted tensors. Our tensor-view platform is designed to allow unrestricted access to tensor memory, while broadcasted tensors were designed for 'read-only' use-cases. Trying to write into a broadcasted tensor needs re-allocation, which Slang is not designed to do.

For now, we enforce contiguity on tensors with any 0 strides.

In the future, we will introduce a ConstTensorView object to allow such tensors to be used as an input.

This patch also propagates name-hint information through structs & arrays of tensors, to allow sensible names for the error messages (before this the error messages were temporary inst numbers, which is nearly impossible to debug)