Add abseil-cpp submodule
This is a dependency of Dawn, which added to address issue #4483.
Dawn, which currently requires abseil-cpp unconditionally, contains abseil-cpp as a
submodule.
However, our CI system by default does not recursively initialize submodules.

Dawn contains lots of submodules which themselves contain submodules, which adds up to
lots of stuff we don't really need, so we don't want to enable recursively initializing
submodules, either.

Thankfully, Dawn allows an external project to specify the directory in which to look for
abseil-cpp, and so we are able to build what we need without recursively initializing
submodules if we have our own abseil-cpp directory to provide to Dawn, which this change
ensures.