Fix the issue in emitFloatCast
In emitFloatCast function, we only considered the input type
is float scalar or float vector, so if the input type is a float
matrix type, it will crash.

We should also handle the float matrix type.

Also, we add some diagnose info to point out the source location
where there is error happened, so in the future it's easier to tell
us what happens.