use `nullptr' for IRStructKey with `IRDerivativeMemberDecoration`
Problem:
Currently, if a decoration refers to its-parent through it operands we will have a recursive dependency and fail. This means storing a `IRStructKey` through a parent field (IRStructField) is fine. It is not fine for a `IRStructKey` to be stored by a `IRStructKey`.

The reason we encounter this problem is because our AutoDiff decoration is not stripped. This is due to the AutoDiff pass not running if we only have a `IRDerivativeMemberDecoration`.

Solution
start AutoDiff pass if we found a `IRDerivativeMemberDecoration`.