Separating slang-test in CI for emit-spirv-via-glsl
When CI test failes, we expect to see the error messages at the end of
each section of CI testing result. But currently when
"emit-spirv-via-glsl" fails, the error messages appear in the middle,
which lead us to believe that there is no errors when you look only at
the end of the result.

This commit separates out the "emit-spirv-via-glsl" case to make the
error message more visiable.