Support an Upper-case variant of [NumThreads]
Closes #4746.

This commit adds a support for "NumThreads" attribute keyword along with
an existing "numthreads", all in lower case.

The attribute keywords in HLSL are case-insensitive.
As an example, one of D3D documents says,
> "The attribute name "Shader" is case insensitive."
https://microsoft.github.io/DirectX-Specs/d3d/WorkGraphs.html

Slang, however, doesn't support the case-insensitivity.
They should be all lower-case or CamelCasing starting with an upper
case.