Add data type of 'oldIndex' to the call info key
In the last commit, we added "oldIndex" directly to the call info
key, this is too conservative, it could generate too many unnecessary
specialized functions, e.g. call(buffer[id]), call(buffer[id + 10])
would generate two same functions with different names.

So instead of doing that, we will use the data type of 'oldIndex' as
call info key.

For the uniformity info, we add a new attribute "IROp_NonUniformAttr",
so we will form a IRAttributedType that encodes both uniformity and data
type, and use it as the key.

Add the modified tests that's missing in last commit.