Combine all tests into one single test
Combine all the tests into one single unit test, and those tests
run will be sub-tests, because when use test-server, each unit test
will be run as a child process, and it will cause race condition
when operating on the 'slang-record' directory as each test will destory
the directory at the end.

Therefore, we make those tests into on unit test, and in this unit
test, we will run each example as a sub-test.

Change the ci to make this test only be run on the gpu-equipped runners,
for other runners we add a white-list file
"expected-failure-record-replay-tests.txt".