fix up wittness table fufillment:
Currently if we have a wrapped type we skip wittness table fufillment checks and try to synth an item incorrectly every-time. This is done to avoid `lookupMember()` checking inheritanceDecl's.

Since 2 levels of inheritance is not very common this caused no issues thus far.

The changes of this PR is to just lookup using `LookupOptions::IgnoreInheritance`