release-note.sh fixes
Add a gh.exe search path that works with cygwin installations.

Do not bail out from the breaking changes loop when a change is
found without PR#. Instead, continue to the next one. This fixes
an issue where [BREAKING] changes appear in the detailed change
list by not in the summary of breaking changes.