Implement texture Gather for WGSL
This commit implements the texture gather functions for WGSL.

The pattern was very similar to how Metal was implemented.
Before copy and paste from the Metal implementation, I had to
clean up the Metal implementation to make it more readable
and maintainable.

Gather functions are available only for 2D and 3D textures.
Their `array` and `depth` variants may or may not be supported depending on the target.
`static_assert` ensures that Gather functions are available only for 2D and 3D textures.

Removed incorrect use of "$p" argument for targeting GLSL.