Enable debug log for VK CTS workflow (#5197)
This commit enables the workflow debug message for running VK CTS
workflow.

Currently three CTS tests are causing a crash of deqp with an exit code
1. And when the debug message printing is enabled, this crash is no
longer happening, which is unexpected. It seems like the issue is on the
CI runner process rather than Slang, driver nor deqp.

We will keep the log message enabled not only to avoid the crash but
also to see more information when things went wrong.

This commit also prints the last 1,000 lines of TestResults.qpa when
deqp had a failing test or deqp was exited unexpectedly. This will help
us to debug the problem faster.