Checkout LLVM as a workflow action rather than git-clone command
Closes #5201

Currently ci.yml builds slang-llvm all the time for all platforms.
Because it takes too long, we store the output files in the cache.
If the cache server is too busy, slang-llvm will be rebuilt, which will take awhile, but it will eventually work out.

The problem is that git-clone was failing for cloning the LLMV repo.
This is a less expected result, because cloning a git repository normally don't fail.
We think that when the git-clone command is manually executed, there might be limit on the network connection based on IP or some sort.
We expect that this problem will be resolved if we use actions/checkout in the workflow.

Also this commit uses a forked repo of llvm under shader-slang
organization to avoid network traffic on the original llvm repo.