Add path to bash to GITHUB_PATH (#5217)
* Add path to bash to GITHUB_PATH

This is a workaround for a problem that keeps reoccurring on one of our self-hosted machines.
Ideally, we should assume that bash shell is available on the runner system, and we should rely on the system-wise setting of PATH to find bash.

This commit is a workaround for a case where bash cannot be found from the runner machine, which shouldn't be a case in a normal situation. It is also not safe to assume that the location of where bash is at a fixed location, because it can vary by how the system is setup.
Once we root cause the issue on the SlangWin4-2, we may remove this workaround later.